LayoutTests/http/tests/referrer-policy/no-referrer/cross-origin-navigation-does-not-inherit-referrer-policy-back-navigation-expected.txt

Tests that referrer policy does not persist across cross-origin navigations with a back navigation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is not ''
PASS successfullyParsed is true

TEST COMPLETE
